+++ src/main/java/mech/mania/engine/game/characters/Monster.java
@@ -24,8 +24,7 @@ public class Monster extends Character {
         if (path.size() < getSpeed()) {
             toMove = path.get(path.size() - 1);
         } else {
-            // TODO remove cast once getSpeed returns int
-            toMove = path.get((int)getSpeed() - 1);
+            toMove = path.get(getSpeed() - 1);
         }
         return toMove;
     }
