+++ src/main/java/mech/mania/engine/game/items/TempStatusModifier.java
@@ -4,7 +4,7 @@ public class TempStatusModifier extends StatusModifier {
     private int duration;
     private double damagePerTurn;               // Concept: poison type of damage that effects player over time
 
-    public TempStatusModifier(double speedChange, double healthChange, double experienceChange,
+    public TempStatusModifier(int speedChange, double healthChange, double experienceChange,
                               double magicDamageChange, double physicalDamageChange, double magicDefenseChange,
                               double physicalDefenseChange, int duration, double damagePerTurn) {
         super(speedChange, healthChange, experienceChange, magicDamageChange, physicalDamageChange, magicDefenseChange,
